Maryland NCES Directory (SY 2003-2024)

Combined data from XLS and XLSX format NCES directory files downloaded from the MSDE Data Downloads website. The school name field has been corrected for consistency and accuracy across years and capitalization of the city and address fields have been standardized (previously used all upper case for both prior to 2010). A grade span value is filled in for older records where grade span was stored in the school type column.

Usage

md_nces_directory

Format

A data frame with 31,579 rows and 14 variables:

year

School year

school_number

School number

school_name

School name

school_type

School type

address

School address

city

City

state

State

zip

Zip code

phone

School phone number

nces_number

NCES Number

create_date

Record creation date

grade_span

Grade span (grade band)

lss_name

LSS name

lss_number

LSS number

Source

MSDE Data Downloads

Details

Last updated 2024-02-22.